Added pet weapon tells (#28)
* Added pet weapon tells

Added tell messages to be sent to the "pet weapons" requester to help them figure out if their pet didn't receive an item. This is just something nice to have until I can figure out how to get fizzles to not break the summon chain.

* Fix Cleric Yaulp

If you have auto-yaulp setup, the cleric was prone to stopping while following the group and getting left behind. This fixes that by adding a check to not cast if moving.