Major Release From Rekka & Ewiclip (#39)
* Fixes to allow large mod shards to be cast without haste aa
Fixes to flush events while waiting for shards to be cast
Deletes all shards before summoning new ones.

* Rekka: Updates for interrupt code for Casting nukes/dots/hots/etc when a tank or xtarget needs a heal. Also fixes for /useitem when an item is used in e3_Casting.

* removed commented if test condition

* Rekka: Loot Only Stackable feature

* Rekka: Fix for useitem in the casting method. Correct return value for items now.

* unsure why this was commented out, as the bard seems to lock up sometimes without this. More testing was needed but so far hasn't hurt but helped. Clicking the epic no longer locks the bard up.

* turn off auto aa

* Merge branch 'master' of https://github.com/Ewiclip/lazarus_mq2_e3

* update rotate in the nuke code to check to see if any other abilities are valid before it forces a rotate.

* forgot to fix echo commands

* bug fix for a echo too close to an if

* Multiple fixes

Spell swap code, MinMana fix, QuickCast text fix.

* Update e3_Assists.inc

Yay for comments

* Rekka:Fix/update for the /GoM flag on buffs. Before it had an issue with spells that had a cooldown, and if was on cooldown it would just 'sit' there.
Now it will correctly skip if on cooldown and use again if a GoM is specified, if not, it will continue with the normal rotation. Probably need to update this for other types (heals/buffs)

* Rekka: Updates to the dot code and fixes for a e3 memory leak.

* Multiple enhancements in description

Removed forgotten echo
Added comments
Minimized the number of times Math.Calc is called
Updated e3_Casting to allow a NoEcho flag to be passed.
Calling e3_cast  instead of /useitem for QuickCast

* Rekka:changed into to bools

* Fixed typo in comment

* Allow spells with a : to queueCast

* INI file name changes

Sets the INI filename to Lazarus if you're logged onto Project Lazarus when  you create a new one. This loads the existing PEQTGC INI even on Lazarus if you already have that and have not created the Lazarus file.

* Placed resistCounter echos behind debug check

* Committing change log

Co-authored-by: Adam Liverman <livermana@gmail.com>