Rekka: heavily modified the bard casting, as to get click issues fixed, bards like normal are special. pause/Resume now work on normal /twist , just do Melody Pause/Melody Resume.
if you cast an item/AA that is greater than cast time, the melody will now also pause/resume where it left off.
Note, this isn't optimal atm until I can get a new player, as it can interrupt the song at 1.5 sec of 3 sec, instead of at the end of the song, so that 1.5 sec has to be redone as the song is resumed.

I hopefully have not broken dynamo...