Rekka: /ImportantSong feature for bard songs. you can have one important song. What will happen if the buff is not found, it will rotate the list to that important song and continue from that point. This is mainly to be 100% sure storm blade never falls off for whatever reason.
Fix issue with buffs when no target was picked. Its a roll back, so you now have to specify the target of a buff period.

Fix issue with mq2linkdb (2.3) causing issues with newer version of mq2 making the game freeze when parsing links.

Fixed issue with invs and the gem lock code not in the correct order, now invs is above gem lock.