Update e3_Sell.inc
Some quality of life features added here that I call flagautosell.  This essentially lets you toggle into a state where whenever you sell an item to the vendor, that item will get updated in the Loot Settings.ini file with  Keep,Sell so that the next time you /autosell at a vendor, that item will get sold automatically.  This eliminates the need to manually edit the Loot Settings.ini file.

Three new INI entries were added with the following default values:

This first setting is the global flag, when this is On, any items you sell to the vendor will get flagged.  When this is Off, items you sell to the vendor will not be sold unless you flip a toggle to enter that state.
Flag Manually Sold Items(On/Off)=Off

The next INI is for a timer.  When the above INI is set to Off and you toggle into a state to flag items sold to the vendor, a timer is set that will automatically disable that state after the timer expires.  This is to help prevent people form inadvertently flagging items to Keep,Sell.
Flag Timer(seconds)=180

This last key is to try and display a warning message when you open a merchant window with a vendor and you're in a state where items sold will get flagged as Keep,Sell.  This will display a message on the screen warning the user that flagautosell is enabled and all items sold will get flagged.   There are some issues with this as not all Merchants say something when you open the merchant window.  Not sure how useful this is if it only works some of the time.
Warn on Sell(On/Off)=On

Several helper slash commands have been added
/printitemstatus - when you have an item on your cursor and you issue this command, it will show you the current Loot Settings.ini value for the key.
/sellthisitem - when you have an item on your cursor and you issue this command, it will flag the item in Loot Settings.ini for Keep,Sell
/keepthisitem - when you have an item on your cursor and you issue this command, it will flag the item in Loot Settings.ini for Keep
/destroythisitem - when you have an item on your cursor and you issue this command, it will flag the item in Loot Settings.ini for Sell
/flagautosell <true/false> - when the default setting is to not flag items, you can call this method with no parms to toggle the state, or pass true/false argument to set the state(true means you'll flag items sold, false means you won't flag items you sell).  This also resets the timer that sets you back to your default state.